+++ transfers/src/androidTest/java/com/example/transfers/ExampleInstrumentedTest.kt
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.transfers", appContext.packageName)
+        assertEquals("com.example.transfers.test", appContext.packageName)
     }
 }