+++ transfers/src/main/java/com/example/transfers/domain/model/Contact.kt
@@ -1,0 +1,14 @@
+package com.example.transfers.domain.model
+
+import java.util.Locale
+
+data class Contact(
+    val firstName: String,
+    val lastName: String,
+    val isClient: Boolean = false
+) {
+    fun contains(searchedText: String): Boolean {
+        return "$firstName $lastName".lowercase(Locale.getDefault())
+            .contains(searchedText.lowercase(Locale.getDefault()))
+    }
+}