+++ transfers/src/main/java/com/example/transfers/domain/repository/ContactsRepository.kt
@@ -1,0 +1,8 @@
+package com.example.transfers.domain.repository
+
+import com.example.transfers.domain.model.Contact
+
+interface ContactsRepository {
+    fun getContacts(): Result<List<Contact>>
+
+}